backend2/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-enumfields==1.0.0
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
jsonfield==2.0.2
Markdown==3.1.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
urllib3==1.25.7
